+++ calico/src/main/scala/calico/html/defs/attrs/AriaAttrs.scala
@@ -1,11 +1,11 @@
 package calico.html.defs.attrs
 
 import calico.html.keys.AriaAttr
-import calico.html.codecs._
+import calico.html.codecs.Codec._; import calico.html.codecs.AsIsCodec._; import calico.html.codecs._
 
 // #NOTE: GENERATED CODE
 //  - This file is generated at compile time from the data in Scala DOM Types
-//  - See `project/DomDefsGenerator.scala` for code generation params
+//  - See `project/src/main/scala/calico/html/codegen/DomDefsGenerator.scala` for code generation params
 //  - Contribute to https://github.com/raquo/scala-dom-types to add missing tags / attrs / props / etc.
 
 trait AriaAttrs[F[_]] {
