+++ calico/src/main/scala/calico/html/defs/eventProps/WindowEventProps.scala
@@ -5,7 +5,7 @@ import org.scalajs.dom
 
 // #NOTE: GENERATED CODE
 //  - This file is generated at compile time from the data in Scala DOM Types
-//  - See `project/DomDefsGenerator.scala` for code generation params
+//  - See `project/src/main/scala/calico/html/codegen/DomDefsGenerator.scala` for code generation params
 //  - Contribute to https://github.com/raquo/scala-dom-types to add missing tags / attrs / props / etc.
 
 /** Window-only Events */
