+++ calico/src/main/scala/calico/html/defs/tags/HtmlTags.scala
@@ -4,7 +4,7 @@ import fs2.dom._
 
 // #NOTE: GENERATED CODE
 //  - This file is generated at compile time from the data in Scala DOM Types
-//  - See `project/DomDefsGenerator.scala` for code generation params
+//  - See `project/src/main/scala/calico/html/codegen/DomDefsGenerator.scala` for code generation params
 //  - Contribute to https://github.com/raquo/scala-dom-types to add missing tags / attrs / props / etc.
 
 trait HtmlTags[F[_], T[_ <: HtmlElement[F]]] {
