+++ project/src/main/scala/calico/html/codegen/DomDefsGenerator.scala
@@ -20,226 +20,6 @@ import com.raquo.domtypes.defs.styles.StyleTraitDefs
 
 object DomDefsGenerator {
 
-  private object generator
-      extends CanonicalGenerator(
-        baseOutputDirectoryPath = "calico/src/main/scala/calico/html",
-        basePackagePath = "calico.html",
-        standardTraitCommentLines = List(
-          "#NOTE: GENERATED CODE",
-          s" - This file is generated at compile time from the data in Scala DOM Types",
-          " - See `project/DomDefsGenerator.scala` for code generation params",
-          " - Contribute to https://github.com/raquo/scala-dom-types to add missing tags / attrs / props / etc."
-        ),
-        format = CodeFormatting()
-      ) {
-
-    override def settersPackagePath: String = basePackagePath + ".modifiers.KeySetter"
-
-    override def scalaJsElementTypeParam: String = "Ref"
-
-    override val baseScalaJsHtmlElementType: String = "HtmlElement[F]"
-
-    override def generateTagsTrait(
-        tagType: TagType,
-        defGroups: List[(String, List[common.TagDef])],
-        printDefGroupComments: Boolean,
-        traitCommentLines: List[String],
-        traitName: String,
-        keyKind: String,
-        baseImplDefComments: List[String],
-        keyImplName: String,
-        defType: DefType): String = {
-      val (defs, defGroupComments) = defsAndGroupComments(defGroups, printDefGroupComments)
-
-      val baseImplDef = if (tagType == HtmlTagType) {
-        List(
-          s"protected def ${keyImplName}[$scalaJsElementTypeParam <: $baseScalaJsHtmlElementType](key: String, void: Boolean = false): ${keyKind}[$scalaJsElementTypeParam]"
-        )
-      } else {
-        List(
-          s"def ${keyImplName}[$scalaJsElementTypeParam <: $baseScalaJsSvgElementType](key: String): ${keyKind}[$scalaJsElementTypeParam] = ${keyKindConstructor(keyKind)}(key)"
-        )
-      }
-
-      val headerLines = List(
-        s"package $tagDefsPackagePath",
-        "",
-        "import fs2.dom._",
-        ""
-      ) ++ standardTraitCommentLines.map("// " + _)
-
-      new TagsTraitGenerator(
-        defs = defs,
-        defGroupComments = defGroupComments,
-        headerLines = headerLines,
-        traitCommentLines = traitCommentLines,
-        traitName = traitName,
-        traitExtends = Nil,
-        traitThisType = None,
-        defType = _ => defType,
-        keyType = tag => keyKind + "[" + TagDefMapper.extractFs2DomElementType(tag) + "]",
-        keyImplName = _ => keyImplName,
-        baseImplDefComments = baseImplDefComments,
-        baseImplDef = baseImplDef,
-        outputImplDefs = true,
-        format = format
-      ).printTrait().getOutput()
-    }
-
-    override def generateAttrsTrait(
-        defGroups: List[(String, List[common.AttrDef])],
-        printDefGroupComments: Boolean,
-        traitCommentLines: List[String],
-        traitName: String,
-        keyKind: String,
-        implNameSuffix: String,
-        baseImplDefComments: List[String],
-        baseImplName: String,
-        namespaceImports: List[String],
-        namespaceImpl: String => String,
-        transformAttrDomName: String => String,
-        defType: DefType): String = {
-      val (defs, defGroupComments) = defsAndGroupComments(defGroups, printDefGroupComments)
-
-      val tagTypes = defs.foldLeft(List[TagType]())((acc, k) => (acc :+ k.tagType).distinct)
-      if (tagTypes.size > 1) {
-        throw new Exception(
-          "Sorry, generateAttrsTrait does not support mixing attrs of different types in one call. You can contribute a PR (please contact us first), or bypass this limitation by calling AttrsTraitGenerator manually.")
-      }
-      val tagType = tagTypes.head
-
-      val baseImplDef = if (tagType == SvgTagType) {
-        List(
-          s"def ${baseImplName}[V](key: String, codec: Codec[V, String], namespace: Option[String]): ${keyKind}[V] = ${keyKindConstructor(keyKind)}(key, codec, namespace)"
-        )
-      } else {
-        List(
-          s"protected def ${baseImplName}[V](key: String, codec: Codec[V, String]): ${keyKind}[F, V] = ${keyKindConstructor(keyKind)}(key, codec)"
-        )
-      }
-
-      val headerLines = List(
-        s"package $attrDefsPackagePath",
-        "",
-        keyTypeImport(keyKind),
-        codecsImport
-      ) ++ namespaceImports ++ List("") ++ standardTraitCommentLines.map("// " + _)
-
-      new AttrsTraitGenerator(
-        defs = defs.map(d => d.copy(domName = transformAttrDomName(d.domName))),
-        defGroupComments = defGroupComments,
-        headerLines = headerLines,
-        traitCommentLines = traitCommentLines,
-        traitName = traitName,
-        traitExtends = Nil,
-        traitThisType = None,
-        defType = _ => defType,
-        keyKind = keyKind,
-        keyImplName = attr => attrImplName(attr.codec, implNameSuffix),
-        baseImplDefComments = baseImplDefComments,
-        baseImplName = baseImplName,
-        baseImplDef = baseImplDef,
-        transformCodecName = _ + "Codec",
-        namespaceImpl = namespaceImpl,
-        outputImplDefs = true,
-        format = format
-      ).printTrait().getOutput()
-    }
-
-    override def generatePropsTrait(
-        defGroups: List[(String, List[common.PropDef])],
-        printDefGroupComments: Boolean,
-        traitCommentLines: List[String],
-        traitName: String,
-        keyKind: String,
-        implNameSuffix: String,
-        baseImplDefComments: List[String],
-        baseImplName: String,
-        defType: DefType): String = {
-
-      val (defs, defGroupComments) = defsAndGroupComments(defGroups, printDefGroupComments)
-
-      val baseImplDef = List(
-        s"def ${baseImplName}[V, DomV](key: String, codec: Codec[V, DomV]): ${keyKind}[F, V, DomV] = ${keyKindConstructor(keyKind)}(key, codec)"
-      )
-
-      val headerLines = List(
-        s"package $propDefsPackagePath",
-        "",
-        keyTypeImport(keyKind),
-        codecsImport,
-        ""
-      ) ++ standardTraitCommentLines.map("// " + _)
-
-      new PropsTraitGenerator(
-        defs = defs,
-        defGroupComments = defGroupComments,
-        headerLines = headerLines,
-        traitCommentLines = traitCommentLines,
-        traitName = traitName,
-        traitExtends = Nil,
-        traitThisType = None,
-        defType = _ => defType,
-        keyKind = keyKind,
-        keyImplName = prop => propImplName(prop.codec, implNameSuffix),
-        baseImplDefComments = baseImplDefComments,
-        baseImplName = baseImplName,
-        baseImplDef = baseImplDef,
-        transformCodecName = _ + "Codec",
-        outputImplDefs = true,
-        format = format
-      ).printTrait().getOutput()
-    }
-
-    override def generateEventPropsTrait(
-        defSources: List[(String, List[common.EventPropDef])],
-        printDefGroupComments: Boolean,
-        traitCommentLines: List[String],
-        traitName: String,
-        traitExtends: List[String],
-        traitThisType: Option[String],
-        baseImplDefComments: List[String],
-        outputBaseImpl: Boolean,
-        keyKind: String,
-        keyImplName: String,
-        defType: DefType): String = {
-      val (defs, defGroupComments) = defsAndGroupComments(defSources, printDefGroupComments)
-
-      val baseImplDef = if (outputBaseImpl)
-        List(
-          s"def ${keyImplName}[Ev <: ${baseScalaJsEventType}](key: String): ${keyKind}[F, Ev] = ${keyKindConstructor(keyKind)}(key)"
-        )
-      else {
-        Nil
-      }
-
-      val headerLines = List(
-        s"package $eventPropDefsPackagePath",
-        "",
-        keyTypeImport(keyKind),
-        scalaJsDomImport,
-        ""
-      ) ++ standardTraitCommentLines.map("// " + _)
-
-      new EventPropsTraitGenerator(
-        defs = defs,
-        defGroupComments = defGroupComments,
-        headerLines = headerLines,
-        traitCommentLines = traitCommentLines,
-        traitName = traitName,
-        traitExtends = traitExtends,
-        traitThisType = traitThisType,
-        defType = _ => defType,
-        keyKind = keyKind,
-        keyImplName = _ => keyImplName,
-        baseImplDefComments = baseImplDefComments,
-        baseImplDef = baseImplDef,
-        outputImplDefs = true,
-        format = format
-      ).printTrait().getOutput()
-    }
-  }
-
   private val cache = new CanonicalCache("project")
 
   def cachedGenerate(): Unit = {
@@ -258,7 +38,7 @@ object DomDefsGenerator {
       val traitName = "HtmlTags"
       val traitNameWithParams = s"$traitName[F[_], T[_ <: HtmlElement[F]]]"
 
-      val fileContent = generator.generateTagsTrait(
+      val fileContent = CalicoGenerator.generateTagsTrait(
         tagType = HtmlTagType,
         // TODO introduce HTMLDialogElement to fs2-dom
         defGroups = defGroups.htmlTagsDefGroups.map {
@@ -283,8 +63,8 @@ object DomDefsGenerator {
         defType = LazyVal
       )
 
-      generator.writeToFile(
-        packagePath = generator.tagDefsPackagePath,
+      CalicoGenerator.writeToFile(
+        packagePath = CalicoGenerator.tagDefsPackagePath,
         fileName = traitName,
         fileContent = fileContent
       )
@@ -295,7 +75,7 @@ object DomDefsGenerator {
     // {
     // val traitName = "SvgTags"
 
-    // val fileContent = generator.generateTagsTrait(
+    // val fileContent = CalicoGenerator.generateTagsTrait(
     // tagType = SvgTagType,
     // defGroups = defGroups.svgTagsDefGroups,
     // printDefGroupComments = false,
@@ -316,8 +96,8 @@ object DomDefsGenerator {
     // defType = LazyVal
     // )
 
-    // generator.writeToFile(
-    // packagePath = generator.tagDefsPackagePath,
+    // CalicoGenerator.writeToFile(
+    // packagePath = CalicoGenerator.tagDefsPackagePath,
     // fileName = traitName,
     // fileContent = fileContent
     // )
@@ -329,7 +109,7 @@ object DomDefsGenerator {
       val traitName = "HtmlAttrs"
       val traitNameWithParams = s"$traitName[F[_]]"
 
-      val fileContent = generator.generateAttrsTrait(
+      val fileContent = CalicoGenerator.generateAttrsTrait(
         defGroups = defGroups.htmlAttrDefGroups.map {
           case (key, vals) =>
             (key, vals.map(attr => attr.copy(scalaValueType = "F, " + attr.scalaValueType)))
@@ -354,8 +134,8 @@ object DomDefsGenerator {
         defType = LazyVal
       )
 
-      generator.writeToFile(
-        packagePath = generator.attrDefsPackagePath,
+      CalicoGenerator.writeToFile(
+        packagePath = CalicoGenerator.attrDefsPackagePath,
         fileName = traitName,
         fileContent = fileContent
       )
@@ -366,7 +146,7 @@ object DomDefsGenerator {
     // {
     // val traitName = "SvgAttrs"
 
-    // val fileContent = generator.generateAttrsTrait(
+    // val fileContent = CalicoGenerator.generateAttrsTrait(
     // defGroups = defGroups.svgAttrDefGroups,
     // printDefGroupComments = false,
     // traitName = traitName,
@@ -388,8 +168,8 @@ object DomDefsGenerator {
     // defType = LazyVal
     // )
 
-    // generator.writeToFile(
-    // packagePath = generator.attrDefsPackagePath,
+    // CalicoGenerator.writeToFile(
+    // packagePath = CalicoGenerator.attrDefsPackagePath,
     // fileName = traitName,
     // fileContent = fileContent
     // )
@@ -409,7 +189,7 @@ object DomDefsGenerator {
         }
       }
 
-      val fileContent = generator.generateAttrsTrait(
+      val fileContent = CalicoGenerator.generateAttrsTrait(
         defGroups = defGroups.ariaAttrDefGroups.map {
           case (key, vals) =>
             (key, vals.map(attr => attr.copy(scalaValueType = "F, " + attr.scalaValueType)))
@@ -434,8 +214,8 @@ object DomDefsGenerator {
         defType = LazyVal
       )
 
-      generator.writeToFile(
-        packagePath = generator.attrDefsPackagePath,
+      CalicoGenerator.writeToFile(
+        packagePath = CalicoGenerator.attrDefsPackagePath,
         fileName = traitName,
         fileContent = fileContent
       )
@@ -447,7 +227,7 @@ object DomDefsGenerator {
       val traitName = "HtmlProps"
       val traitNameWithParams = s"$traitName[F[_]]"
 
-      val fileContent = generator.generatePropsTrait(
+      val fileContent = CalicoGenerator.generatePropsTrait(
         defGroups = defGroups.propDefGroups.map {
           case (key, vals) =>
             (key, vals.map(attr => attr.copy(scalaValueType = "F, " + attr.scalaValueType)))
@@ -470,8 +250,8 @@ object DomDefsGenerator {
         defType = LazyVal
       )
 
-      generator.writeToFile(
-        packagePath = generator.propDefsPackagePath,
+      CalicoGenerator.writeToFile(
+        packagePath = CalicoGenerator.propDefsPackagePath,
         fileName = traitName,
         fileContent = fileContent
       )
@@ -489,7 +269,7 @@ object DomDefsGenerator {
       )
 
       {
-        val fileContent = generator.generateEventPropsTrait(
+        val fileContent = CalicoGenerator.generateEventPropsTrait(
           defSources = defGroups.globalEventPropDefGroups.map {
             case (key, vals) =>
               (
@@ -514,8 +294,8 @@ object DomDefsGenerator {
           defType = LazyVal
         )
 
-        generator.writeToFile(
-          packagePath = generator.eventPropDefsPackagePath,
+        CalicoGenerator.writeToFile(
+          packagePath = CalicoGenerator.eventPropDefsPackagePath,
           fileName = baseTraitName,
           fileContent = fileContent
         )
@@ -523,7 +303,7 @@ object DomDefsGenerator {
 
       subTraits.foreach {
         case (traitName, traitNameWithParams, eventPropsDefGroups) =>
-          val fileContent = generator.generateEventPropsTrait(
+          val fileContent = CalicoGenerator.generateEventPropsTrait(
             defSources = eventPropsDefGroups.map {
               case (key, vals) =>
                 (
@@ -542,8 +322,8 @@ object DomDefsGenerator {
             defType = LazyVal
           )
 
-          generator.writeToFile(
-            packagePath = generator.eventPropDefsPackagePath,
+          CalicoGenerator.writeToFile(
+            packagePath = CalicoGenerator.eventPropDefsPackagePath,
             fileName = traitName,
             fileContent = fileContent
           )
@@ -555,7 +335,7 @@ object DomDefsGenerator {
     // {
     // val traitName = "StyleProps"
 
-    // val fileContent = generator.generateStylePropsTrait(
+    // val fileContent = CalicoGenerator.generateStylePropsTrait(
     // defSources = defGroups.stylePropDefGroups,
     // printDefGroupComments = true,
     // traitCommentLines = Nil,
@@ -581,8 +361,8 @@ object DomDefsGenerator {
     // outputUnitTraits = true
     // )
 
-    // generator.writeToFile(
-    // packagePath = generator.stylePropDefsPackagePath,
+    // CalicoGenerator.writeToFile(
+    // packagePath = CalicoGenerator.stylePropDefsPackagePath,
     // fileName = traitName,
     // fileContent = fileContent
     // )
@@ -592,7 +372,7 @@ object DomDefsGenerator {
 
     // {
     // StyleTraitDefs.defs.foreach { styleTrait =>
-    // val fileContent = generator.generateStyleKeywordsTrait(
+    // val fileContent = CalicoGenerator.generateStyleKeywordsTrait(
     // defSources = styleTrait.keywordDefGroups,
     // printDefGroupComments = styleTrait.keywordDefGroups.length > 1,
     // traitCommentLines = Nil,
@@ -608,8 +388,8 @@ object DomDefsGenerator {
     // allowSuperCallInOverride = false // can't access lazy val from `super`
     // )
 
-    // generator.writeToFile(
-    // packagePath = generator.styleTraitsPackagePath(),
+    // CalicoGenerator.writeToFile(
+    // packagePath = CalicoGenerator.styleTraitsPackagePath(),
     // fileName = styleTrait.scalaName.replace("[_]", ""),
     // fileContent = fileContent
     // )
